Fixed/ignored warnings in Libsass

+++ LibSass/src/util.cpp
@@ -69,7 +69,7 @@ namespace Sass {
 
           // convert the extracted hex string to code point value
           // ToDo: Maybe we could do this without creating a substring
-          uint32_t cp = strtol(s.substr (i + 1, len - 1).c_str(), nullptr, 16);
+          uint32_t cp = (uint32_t)strtol(s.substr (i + 1, len - 1).c_str(), nullptr, 16);
 
           if (cp == 0) cp = 0xFFFD;
 
@@ -361,7 +361,7 @@ namespace Sass {
 
           // convert the extracted hex string to code point value
           // ToDo: Maybe we could do this without creating a substring
-          uint32_t cp = strtol(s.substr (i + 1, len - 1).c_str(), nullptr, 16);
+          uint32_t cp = (uint32_t)strtol(s.substr (i + 1, len - 1).c_str(), nullptr, 16);
 
           // assert invalid code points
           if (cp == 0) cp = 0xFFFD;
